Fix for broken stdout and file output
The sym application fails on stdout or the output switch to a file:

```
$ sym -h

/usr/local/opt/rbenv/versions/2.4.0/lib/ruby/gems/2.4.0/gems/sym-2.2.1/lib/sym/app/cli.rb:112:in `select_output_stream': Can not determine output class from arguments {:encrypt=>false, :decrypt=>false, :edit=>false, :generate=>false, :password=>false, :private_key=>nil, :keyfile=>nil, :keychain=>nil, :interactive=>false, :cache_password=>false, :cache_for=>nil, :cache_provider=>nil, :string=>nil, :file=>nil, :output=>nil, :backup=>false, :verbose=>false, :trace=>false, :debug=>false, :quiet=>false, :version=>false, :no_color=>false, :no_environment=>false, :bash_completion=>nil, :examples=>false, :help=>true} (RuntimeError)
```

The quiet switch succeeds. This PR fixes the loading of the output classes
so that stdout and writing to a file work. (I still have no idea why using
`require` instead of `require_relative` results in this error.)